+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [Unreleased]
+
+- [TD-7696]:
+  - BG:
+    - Access to permissions model
+    - Interact with service search query
+    - Fetch store information
+  - DD:
+    - Interact with service search query
+  - LM:
+   - Refresh link search data
+   - Allow to delete stale relations
+   - Count relations given a set of params
+  - QX:
+    - Fetch store information
+
 ## [7.14.1] 2025-12-09
 
 - [TD-7678] Update TdBg functions for search concepts by name, domain and template